initial version of the snoods client/server
For the sake of simplicity, the same script runs either
the client or the server, depending on the commandline.
This is awkward because it requires the server to have the
Tkinter libraries installed, which is an annoying dependency
if the server runs on a headless server.  This might change
in the future.