wip: add board_id
work in progress because this breaks the ability to load
messages from a file: this worked when there was only one
board (so there was no question which board the messages
should go to).  Now that we've added the ability to add
new boards on the fly, it's not obvious how this should work.
(this functionality might need to be pruned)