fix new join request; add response
this is still awkward because when the client first joins the
server, it connects, by default, to the default board and so
the first thing it gets is all the updates for the default
whiteboard, and *then* the new board response, followed by
all of the updates for the new board.  Not a problem if there's
nothing on the default board, but annoying if there is.