fix the logic for joining a board
if we're joining the default board, then we don't
need to send a join message.  Similarly, if we've
sent a join message, we can ignore updates until
the expected board name is the same as the one returned
to us (so we can ignore spurious responses from joining
the "default" board).  The logic is slightly redundant
here (we're adding more checks than needed), but that's
better than the alternative